Move xhr.js to util/
