+++ lib/web.js
@@ -8,7 +8,7 @@ var graphUtil = require('./util/graph-util')
 var SolidResponse = require('./solid/response')
 var SolidContainer = require('./solid/container')
 var Vocab = require('./vocab')
-var XMLHttpRequest = require('./xhr')
+var XMLHttpRequest = require('./util/xhr')
 
 /**
  * Provides a collection of Solid/LDP web operations (CRUD)
